查看编辑攻略等

+++ webserver/src/main/java/com/truemen/api/guide/repository/GuideRepository.java
@@ -1,9 +1,10 @@
 package com.truemen.api.guide.repository;
 
-import com.truemen.api.guide.model.Guide;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.truemen.api.guide.model.Guide;
+
 @Repository
-public interface GuideRepository extends JpaRepository<Guide, Integer> {
+public interface GuideRepository extends JpaRepository<Guide, Long> {
 }
